requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
openai==2.30.0
requests==2.33.0
openenv-core==0.2.3
python-dotenv==1.2.2
jinja2==3.1.6